:zap: policies done

+++ routes/api.php
@@ -26,49 +26,49 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 //Auth routes for Sanctum
-Route::post('/login', [AuthController::class, 'login'])->name('login');
+Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
-Route::get('/me', [AuthController::class, 'me'])->middleware('auth:sanctum')->name('me');
-Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum')->name('logout');
+Route::get('/me', [AuthController::class, 'me'])->middleware('auth:sanctum');                            // auth
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');                    // auth
 
 //All routes for UserController
-Route::get('/users', [UserController::class, 'index'])->middleware('auth:sanctum');
-Route::get('/user/{id}', [UserController::class, 'show'])->middleware('auth:sanctum');
-Route::post('/user', [UserController::class, 'store'])->middleware('auth:sanctum');
-Route::delete('/user/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
-Route::put('/user/{id}', [UserController::class, 'update'])->middleware('auth:sanctum');
+Route::get('/users', [UserController::class, 'index']);
+Route::get('/user/{id}', [UserController::class, 'show']);
+Route::post('/user', [UserController::class, 'store'])->middleware(['auth:sanctum', 'admin']);                      // auth && admin
+Route::delete('/user/{id}', [UserController::class, 'destroy'])->middleware('auth:sanctum');             // auth && (admin || user_id == id)
+Route::put('/user/{id}', [UserController::class, 'update'])->middleware('auth:sanctum');                 // auth && (admin || user_id == id)
 
 //All routes for EventController
 Route::get('/events', [EventController::class, 'index']);
 Route::get('/event/{id}', [EventController::class, 'show']);
-Route::post('/event', [EventController::class, 'store']);
-Route::delete('/event/{id}', [EventController::class, 'destroy']);
-Route::put('/event/{id}', [EventController::class, 'update']);
+Route::post('/event', [EventController::class, 'store'])->middleware(['auth:sanctum']);                            // auth
+Route::delete('/event/{id}', [EventController::class, 'delete'])->middleware(['auth:sanctum']);                    // auth && (admin || creator_id == user_id)
+Route::put('/event/{id}', [EventController::class, 'update'])->middleware('auth:sanctum');              // auth && (admin || creator_id == user_id)
 
 //All routes for DrinkController
 Route::get('/drinks', [DrinkController::class, 'index']);
 Route::get('/drink/{id}', [DrinkController::class, 'show']);
-Route::post('/drink', [DrinkController::class, 'store']);
-Route::delete('/drink/{id}', [DrinkController::class, 'destroy']);
-Route::put('/drink/{id}', [DrinkController::class, 'update']);
+Route::post('/drink', [DrinkController::class, 'store'])->middleware(['auth:sanctum', 'admin']);                    // auth && admin
+Route::delete('/drink/{id}', [DrinkController::class, 'destroy'])->middleware(['auth:sanctum', 'admin']);           // auth && admin
+Route::put('/drink/{id}', [DrinkController::class, 'update'])->middleware(['auth:sanctum', 'admin']);               // auth && admin
 
 //All routes for Role
 Route::get('/roles', [RoleController::class, 'index']);
 Route::get('/role/{id}', [RoleController::class, 'show']);
-Route::post('/role', [RoleController::class, 'store']);
-Route::delete('/role/{id}', [RoleController::class, 'destroy']);
-Route::put('/role/{id}', [RoleController::class, 'update']);
+Route::post('/role', [RoleController::class, 'store'])->middleware(['auth:sanctum', 'admin']);                      // auth && admin
+Route::delete('/role/{id}', [RoleController::class, 'destroy'])->middleware(['auth:sanctum', 'admin']);             // auth && admin
+Route::put('/role/{id}', [RoleController::class, 'update'])->middleware(['auth:sanctum', 'admin']);                 // auth && admin
 
 //All routes for Category
 Route::get('/categories', [CategoryController::class, 'index']);
 Route::get('/category/{id}', [CategoryController::class, 'show']);
-Route::post('/category', [CategoryController::class, 'store']);
-Route::delete('/category/{id}', [CategoryController::class, 'destroy']);
-Route::put('/category/{id}', [CategoryController::class, 'update']);
+Route::post('/category', [CategoryController::class, 'store'])->middleware(['auth:sanctum', 'admin']);              // auth && admin
+Route::delete('/category/{id}', [CategoryController::class, 'destroy'])->middleware(['auth:sanctum', 'admin']);     // auth && admin
+Route::put('/category/{id}', [CategoryController::class, 'update'])->middleware(['auth:sanctum', 'admin']);         // auth && admin
 
 //All routes for rank
 Route::get('/ranks', [RankController::class, 'index']);
 Route::get('/rank/{id}', [RankController::class, 'show']);
-Route::post('/rank', [RankController::class, 'store']);
-Route::delete('/rank/{id}', [RankController::class, 'destroy']);
-Route::put('/rank/{id}', [RankController::class, 'update']);
+Route::post('/rank', [RankController::class, 'store'])->middleware(['auth:sanctum', 'admin']);                      // auth && admin
+Route::delete('/rank/{id}', [RankController::class, 'destroy'])->middleware(['auth:sanctum', 'admin']);             // auth && admin
+Route::put('/rank/{id}', [RankController::class, 'update'])->middleware(['auth:sanctum', 'admin']);                 // auth && admin
